Add in memory database for dev env.

+++ src/services/product/src/main/java/com/microshop/microservice/product/controller/ProductController.java
@@ -2,6 +2,8 @@ package com.microshop.microservice.product.controller;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.microshop.microservice.product.model.Product;
@@ -17,34 +19,49 @@ public class ProductController {
     ProductService productService;
 
     @GetMapping
-    public List<Product> list() {
-        return productService.list();
+    public ResponseEntity<List<Product>> list() {
+        return new ResponseEntity<List<Product>>(productService.list(), HttpStatus.OK);
     }
 
     @GetMapping(value = "{productId}")
-    public Product get(@PathVariable Integer productId) {
+    public ResponseEntity<Product> get(@PathVariable Integer productId) {
         var product = productService.get(productId);
-        return product;
+        if (product == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<Product>(product, HttpStatus.OK);
     }
 
-    @GetMapping(value = "{productName}")
-    public List<Product> get(@PathVariable String productName) {
-        var products = productService.get(productName);
-        return products;
-    }
+//    @GetMapping(value = "{productName}")
+//    public List<Product> get(@PathVariable String productName) {
+//        var products = productService.get(productName);
+//        return products;
+//    }
 
     @PostMapping
-    public Product create(@RequestBody final Product product) {
+    public Product create(@RequestBody final Product product) {#
         return productService.create(product);
     }
 
     @DeleteMapping(value = "{productId}")
     public void delete(@PathVariable Integer productId) {
+        if (productService.get(productId) == null) {
+            throw new RuntimeException("Product not found");
+        }
         productService.delete(productId);
     }
 
     @RequestMapping(value = "{productId}", method = RequestMethod.PUT)
     public Product update(@PathVariable Integer productId, @RequestBody Product product) {
+
+        if (productService.get(productId) == null) {
+            throw new RuntimeException("Product not found");
+        }
+
+        if (product.getProductId() != null && product.getProductId() != productId) {
+            throw new RuntimeException("Product id doesn't match");
+        }
+
         Product existingProduct = productService.get(productId);
         BeanUtils.copyProperties(product, existingProduct, "product_id");
         return productService.update(existingProduct);
